test: seed persisted workbench state before Browser E2E
