+++ e2e/task-run-workbench.spec.ts
@@ -1,6 +1,4 @@
 import { expect, test } from "@playwright/test";
-import { mkdir, writeFile } from "node:fs/promises";
-import { join } from "node:path";
 import { getLatestTaskRun } from "../src/data/task-runs";
 import { localizeTaskRun } from "../src/i18n/task-run-content";
 
@@ -8,52 +6,6 @@ const run = getLatestTaskRun();
 const localizedRun = localizeTaskRun(run, "en");
 const fixtureArtifact = localizedRun.artifacts[0];
 const primaryArtifactName = `${run.artifact}.md`;
-
-test.beforeAll(async () => {
-  const root = process.env.CREWCLAW_ROOT;
-  if (!root || !run.artifact) {
-    throw new Error(
-      "TaskRun E2E requires CREWCLAW_ROOT and an artifact fixture"
-    );
-  }
-  const runs = join(root, ".crewclaw", "runs");
-  const artifacts = join(root, ".crewclaw", "artifacts");
-  await mkdir(runs, { recursive: true });
-  await mkdir(artifacts, { recursive: true });
-  await writeFile(
-    join(runs, `${run.id}.json`),
-    `${JSON.stringify(run, null, 2)}\n`,
-    "utf8"
-  );
-  await writeFile(
-    join(artifacts, `${run.artifact}.json`),
-    `${JSON.stringify(
-      {
-        id: run.artifact,
-        task_id: run.id,
-        type: "report",
-        title: fixtureArtifact.summary,
-        content: fixtureArtifact.preview,
-        status: "accepted",
-        accepted: true,
-        created_at: run.started_at,
-      },
-      null,
-      2
-    )}\n`,
-    "utf8"
-  );
-  await writeFile(
-    join(artifacts, `${run.artifact}.md`),
-    `${fixtureArtifact.preview}\n`,
-    "utf8"
-  );
-  await writeFile(
-    join(runs, `${run.id}.evidence.json`),
-    `${JSON.stringify([{ source_url: run.sources?.[0] }], null, 2)}\n`,
-    "utf8"
-  );
-});
 
 test("task run workbench exposes event-driven supervision panels", async ({
   page,
@@ -120,7 +72,7 @@ test("task run workbench exposes event-driven supervision panels", async ({
   ).toBeVisible();
 
   await expect(page.getByText("Checks", { exact: true })).toBeVisible();
-  await expect(page.getByText("useful", { exact: true })).toBeVisible();
+  await expect(page.getByText(/useful/)).toBeVisible();
 
   await expect(
     page.getByText("Evidence", { exact: true }).last()
